requirements_initial.txt
pillow
cookiecutter

django_tables2
django-crispy-forms
django-bootstrap4
django-colorful
django-colorfield
django-phone-field
django-countries
django-widget-tweaks
django-summernote
django-bootstrap-modal-forms
django-autocomplete-light

postgres
py-postgresql

djangorestframework
markdown
django-filter

django-sitetree

argon2-cffi
bcrypt
pbkdf2
